// TvServer.WebFrontend/src/apis/TvControlClient.ts
export enum DeviceStatus {
    _0 = 0,
    _1 = 1,
    _2 = 2,
}

export class RokuApp implements IRokuApp {
    id?: string | undefined;
    type?: string | undefined;
    version?: string | undefined;
    name?: string | undefined;

    constructor(data?: IRokuApp) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.type = _data["type"];
            this.version = _data["version"];
            this.name = _data["name"];
        }
    }

    static fromJS(data: any): RokuApp {
        data = typeof data === 'object' ? data : {};
        let result = new RokuApp();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["type"] = this.type;
        data["version"] = this.version;
        data["name"] = this.name;
        return data;
    }
}

export interface IRokuApp {
    id?: string | undefined;
    type?: string | undefined;
    version?: string | undefined;
    name?: string | undefined;
}

export class RokuApps implements IRokuApps {
    id?: string | undefined;
    appList?: RokuApp[] | undefined;

    constructor(data?: IRokuApps) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            if (Array.isArray(_data["appList"])) {
                this.appList = [] as any;
                for (let item of _data["appList"])
                    this.appList!.push(RokuApp.fromJS(item));
            }
        }
    }

    static fromJS(data: any): RokuApps {
        data = typeof data === 'object' ? data : {};
        let result = new RokuApps();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        if (Array.isArray(this.appList)) {
            data["appList"] = [];
            for (let item of this.appList)
                data["appList"].push(item ? item.toJSON() : undefined as any);
        }
        return data;
    }
}

export interface IRokuApps {
    id?: string | undefined;
    appList?: RokuApp[] | undefined;
}
// ...
export class RokuTvDto implements IRokuTvDto {
    id!: string;
    deviceName!: string;
    lastKnownIp?: string | undefined;
    status?: DeviceStatus;
    rokuApps?: RokuApps;
    rokuDeviceInfo?: RokuDeviceInfo;

    constructor(data?: IRokuTvDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.deviceName = _data["deviceName"];
            this.lastKnownIp = _data["lastKnownIp"];
            this.status = _data["status"];
            this.rokuApps = _data["rokuApps"] ? RokuApps.fromJS(_data["rokuApps"]) : undefined as any;
            this.rokuDeviceInfo = _data["rokuDeviceInfo"] ? RokuDeviceInfo.fromJS(_data["rokuDeviceInfo"]) : undefined as any;
        }
    }

    static fromJS(data: any): RokuTvDto {
        data = typeof data === 'object' ? data : {};
        let result = new RokuTvDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["deviceName"] = this.deviceName;
        data["lastKnownIp"] = this.lastKnownIp;
        data["status"] = this.status;
        data["rokuApps"] = this.rokuApps ? this.rokuApps.toJSON() : undefined as any;
        data["rokuDeviceInfo"] = this.rokuDeviceInfo ? this.rokuDeviceInfo.toJSON() : undefined as any;
        return data;
    }
}

export interface IRokuTvDto {
    id: string;
    deviceName: string;
    lastKnownIp?: string | undefined;
    status?: DeviceStatus;
    rokuApps?: RokuApps;
    rokuDeviceInfo?: RokuDeviceInfo;
}
// ...
export class SamsungAppData implements ISamsungAppData {
    id?: string | undefined;
    appId?: string | undefined;
    appType?: number;
    icon?: string | undefined;
    isLock?: number;
    name?: string | undefined;

    constructor(data?: ISamsungAppData) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.appId = _data["appId"];
            this.appType = _data["appType"];
            this.icon = _data["icon"];
            this.isLock = _data["isLock"];
            this.name = _data["name"];
        }
    }

    static fromJS(data: any): SamsungAppData {
        data = typeof data === 'object' ? data : {};
        let result = new SamsungAppData();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["appId"] = this.appId;
        data["appType"] = this.appType;
        data["icon"] = this.icon;
        data["isLock"] = this.isLock;
        data["name"] = this.name;
        return data;
    }
}

export interface ISamsungAppData {
    id?: string | undefined;
    appId?: string | undefined;
    appType?: number;
    icon?: string | undefined;
    isLock?: number;
    name?: string | undefined;
}
// ...
export class SamsungTvDto implements ISamsungTvDto {
    id!: string;
    deviceName!: string;
    lastKnownIp?: string | undefined;
    status?: DeviceStatus;
    samsungApps?: SamsungAppData[] | undefined;
    deviceInfo?: SamsungTvInfo;

    constructor(data?: ISamsungTvDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.deviceName = _data["deviceName"];
            this.lastKnownIp = _data["lastKnownIp"];
            this.status = _data["status"];
            if (Array.isArray(_data["samsungApps"])) {
                this.samsungApps = [] as any;
                for (let item of _data["samsungApps"])
                    this.samsungApps!.push(SamsungAppData.fromJS(item));
            }
            this.deviceInfo = _data["deviceInfo"] ? SamsungTvInfo.fromJS(_data["deviceInfo"]) : undefined as any;
        }
    }

    static fromJS(data: any): SamsungTvDto {
        data = typeof data === 'object' ? data : {};
        let result = new SamsungTvDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["deviceName"] = this.deviceName;
        data["lastKnownIp"] = this.lastKnownIp;
        data["status"] = this.status;
        if (Array.isArray(this.samsungApps)) {
            data["samsungApps"] = [];
            for (let item of this.samsungApps)
                data["samsungApps"].push(item ? item.toJSON() : undefined as any);
        }
        data["deviceInfo"] = this.deviceInfo ? this.deviceInfo.toJSON() : undefined as any;
        return data;
    }
}

export interface ISamsungTvDto {
    id: string;
    deviceName: string;
    lastKnownIp?: string | undefined;
    status?: DeviceStatus;
    samsungApps?: SamsungAppData[] | undefined;
    deviceInfo?: SamsungTvInfo;
}

export class SamsungTvInfo implements ISamsungTvInfo {
    device?: SamsungAdditionalDeviceInfo;
    id?: string | undefined;
    isSupport?: string | undefined;
    name?: string | undefined;
    remote?: string | undefined;
    type?: string | undefined;
    uri?: string | undefined;
    version?: string | undefined;

    constructor(data?: ISamsungTvInfo) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.device = _data["device"] ? SamsungAdditionalDeviceInfo.fromJS(_data["device"]) : undefined as any;
            this.id = _data["id"];
            this.isSupport = _data["isSupport"];
            this.name = _data["name"];
            this.remote = _data["remote"];
            this.type = _data["type"];
            this.uri = _data["uri"];
            this.version = _data["version"];
        }
    }

    static fromJS(data: any): SamsungTvInfo {
        data = typeof data === 'object' ? data : {};
        let result = new SamsungTvInfo();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["device"] = this.device ? this.device.toJSON() : undefined as any;
        data["id"] = this.id;
        data["isSupport"] = this.isSupport;
        data["name"] = this.name;
        data["remote"] = this.remote;
        data["type"] = this.type;
        data["uri"] = this.uri;
        data["version"] = this.version;
        return data;
    }
}

export interface ISamsungTvInfo {
    device?: SamsungAdditionalDeviceInfo;
    id?: string | undefined;
    isSupport?: string | undefined;
    name?: string | undefined;
    remote?: string | undefined;
    type?: string | undefined;
    uri?: string | undefined;
    version?: string | undefined;
}